setting file structure
