+++ src/App.js
@@ -1,10 +1,9 @@
 import React from 'react';
-import './style.css';
+import '../src/style/style.css';
 import { Provider } from 'react-redux'
 import store from './redux/store'
-import CakeContainer from './components/cakeContainer';
-import HooksCakeContainer from './components/HooksCakeContainer';
-import IceCreamContainer from './components/iceCreamContainer';
+import CakeContainer from './components/CakeContainer';
+import IceCreamContainer from './components/IceCreamContainer';
 import NewCakeContainer from './components/NewCakeContainer'
 import ItemContainer from './components/ItemContainer';
 import UserContainer from './components/UserContainer';
@@ -13,14 +12,12 @@ function App() {
   return (
     <Provider store = {store}>
       <div className="App">
-      <UserContainer />
-      {/*
-      <ItemContainer cake/>
-      <ItemContainer />
-        <HooksCakeContainer />
+        <UserContainer />
+        <ItemContainer cake/>
+        <ItemContainer />
         <CakeContainer />
         <IceCreamContainer />
-        <NewCakeContainer />*/}
+        <NewCakeContainer />
       </div>
     </Provider>
   );
